requirements.txt
attrs==23.2.0
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
Flask==3.0.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
PyJWT==2.8.0
referencing==0.34.0
rpds-py==0.18.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
Werkzeug==3.0.1
